+++ creusot/src/backend.rs
@@ -1,2 +1,3 @@
 pub(crate) mod logic;
+pub(crate) mod program;
 pub(crate) mod term;
